backend/requirements.txt
fastapi==0.117.1
uvicorn==0.36.0
crewai==0.193.2
python-socketio==5.13.0
aiofiles==24.1.0
PyYAML==6.0.2
watchdog==6.0.0
python-multipart==0.0.20
psutil==5.9.5
python-dotenv==1.0.0
